Remove paralel term parsing
In order to compute term in parallel we consider only case
when we receive closure from stack and term of that closure is a variable.
Than if this variable is not in environment of that closure
we can create n processes where n is size of stack and compute each
closure in parallel. Than in the process we can meet this pattern
again, so we cannot analyze this during parsing. This should be solved
in the runtime.